# Performance optimization for SVM predict