--- a/cpp/daal/src/threading/threading.h
+++ b/cpp/daal/src/threading/threading.h
@@ -264,8 +264,9 @@ class ls : public tlsBase
 {
 public:
     template <typename lambdaType>
-    explicit ls(const lambdaType & lambda)
+    explicit ls(const lambdaType & lambda, const bool isTls = false)
     {
+        _isTls              = isTls;
         lambdaType * locall = new lambdaType(lambda);
         d                   = new tls_deleter_<lambdaType>();
 
@@ -274,36 +275,40 @@ public:
         void * a        = const_cast<void *>(ac);
         voidLambda      = a;
 
-        lsPtr = _daal_get_ls_ptr(a, tls_func<lambdaType>);
+        lsPtr = _isTls ? _daal_get_tls_ptr(a, tls_func<lambdaType>) : _daal_get_ls_ptr(a, tls_func<lambdaType>);
     }
 
     virtual ~ls()
     {
         d->del(voidLambda);
         delete d;
-        _daal_del_ls_ptr(lsPtr);
+        _isTls ? _daal_del_tls_ptr(lsPtr) : _daal_del_ls_ptr(lsPtr);
     }
 
     F local()
     {
-        void * pf = _daal_get_ls_local(lsPtr);
+        void * pf = _isTls ? _daal_get_tls_local(lsPtr) : _daal_get_ls_local(lsPtr);
         return (static_cast<F>(pf));
     }
 
-    void release(F p) { _daal_release_ls_local(lsPtr, p); }
+    void release(F p)
+    {
+        if (!_isTls) _daal_release_ls_local(lsPtr, p);
+    }
 
     template <typename lambdaType>
     void reduce(const lambdaType & lambda)
     {
         const void * ac = static_cast<const void *>(&lambda);
         void * a        = const_cast<void *>(ac);
-        _daal_reduce_ls(lsPtr, a, tls_reduce_func<F, lambdaType>);
+        _isTls ? _daal_reduce_tls(lsPtr, a, tls_reduce_func<F, lambdaType>) : _daal_reduce_ls(lsPtr, a, tls_reduce_func<F, lambdaType>);
     }
 
 private:
     void * lsPtr;
     void * voidLambda;
     tls_deleter * d;
+    bool _isTls;
 };
 
 template <typename F>