1. Improve blocking for SVM predict by splitting support vectors
2. Extend LS storage memory struct for TLS case

Problem | Before, ms | After, ms | Speedup
--| -- | -- | --
ijcnn | 940 | 110 | **8.2**
klaverjas | 7400 | 2030 | **2.6**
connect4  | 1050 | 340 | **3.2**
mnist | 3120 | 1680 | **1.9**
sensit | 1400 | 400 | **3.5**
skin_segmentation | 2540 | 920 | **2.8**
covertype| 32400 | 11700 | **2.8**